// equipoDeFutbol/src/app/interfaces/jugadores.ts
export interface Jugadores{
    $key?:string;
    nombre:string;
    apellido:string;
    posicion: NumeroPosicion;
    peso:number;
    altura:number;
    nacionalidad:Countries;
    izquierdo:boolean;
}

export enum NumeroPosicion{
    arquero = 1,
    defensaCentral = 3,
    lateralIzquierdo = 6,
    lateralDerecho = 7,
    volanteCreativo = 10,
    delantero = 9
}